added some new classes one lev shop

+++ OOP/08-OOP-in-Java/src/oneLvShop/Buyable.java
@@ -1,7 +1,5 @@
 package oneLvShop;
 
-import java.math.BigDecimal;
-
 public interface Buyable {
-	BigDecimal getPrice();
+	double getPrice();
 }
